refactor : #67 id 타입 변경에 따른 side effect 해결

+++ src/main/java/com/example/team1_be/domain/Apply/ApplyRepository.java
@@ -2,5 +2,5 @@ package com.example.team1_be.domain.Apply;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplyRepository extends JpaRepository<Apply, Integer> {
+public interface ApplyRepository extends JpaRepository<Apply, Long> {
 }
